challenge5/requirements.txt
Flask
Jinja2
Werkzeug
certifi
chardet
gunicorn
requests
tensorflow
keras
numpy
pillow
